refactor(engine): drop gitnexus cypher fallback, native lbug only
The cypher fallback was the v1 loader kept alive after native lbug became
the primary, pinned-in-lockstep path. It only rescued an explicitly
unsupported config (newer-than-pinned gitnexus), was never exercised in
CI, and carried ~570 lines of subprocess/markdown/paging/parallel code.

- Delete gitnexus_cypher adapter + mdg/cypher loader + bench harness.
- from_lbug_path: dir -> JSON, file -> native lbug, else NotFound (drops
  the now-unused project_root/branch args).
- Remove CypherFailed/CypherUnavailable MdgError variants.

cargo build/clippy -D warnings/fmt/mdg tests all green.
