+++ topos/engine/src/graphs/mdg/mod.rs
@@ -1,7 +1,6 @@
 //! The Module Dependency Graph — the inter-module (import/call/inherit)
 //! view parsed from GitNexus output, feeding the COMPOSABLE generator.
 
-mod cypher;
 mod json;
 mod ladybug;
 pub mod models;
